renamed stuff in steps keras, added clear session
