+++ config.yaml
@@ -5,12 +5,15 @@ name: minerva
 parameters:
 
   # Local setup
-  #data_dir: resources/whales/data # for instance resources/whales/data
-  #solution_dir:  output/path_to_your_solution # for instance /output/resources/whales/solution/localization
+#  data_dir: resources/whales/data # for instance resources/whales/data
+#  solution_dir:  output/path_to_your_solution # for instance /output/resources/whales/solution/localization
+
+  data_dir: /mnt/ml-team/minerva/whales/data
+  solution_dir:  /mnt/ml-team/minerva/resources/fashion_mnist/solution
 
   # Cloud setup
-  data_dir: /public/whales # for instance /public/whales
-  solution_dir: /output/path_to_your_solution # /output/path_to_your_solution
+#  data_dir: /public/whales # for instance /public/whales
+#  solution_dir: /output/path_to_your_solution # /output/path_to_your_solution
 
 exclude:
   - resources
@@ -22,4 +25,4 @@ exclude:
   - .ipynb_checkpoints
 
 # Comment if local
-pip-requirements-file: requirements.txt
+#pip-requirements-file: requirements.txt
