+++ minerva/fashion_mnist/config.py
@@ -50,7 +50,7 @@ SOLUTION_CONFIG = {
                                                            'nesterov': True
                                                            },
                                       },
-              'training_config': {'epochs': 200},
+              'training_config': {'epochs': 2},
               'callbacks_config': {'patience': 10,
                                    'model_name': 'simplenet'}
               },
